requirements.txt
streamlit
pandas
pymupdf
sentence-transformers
faiss-cpu
rank-bm25
numpy
rank-bm25
huggingface_hub
python-dotenv